configure signup/login logic for auth server for day 16

+++ day16/index.js
@@ -9,13 +9,48 @@ app.use('/', express.static('public'));
 
 app.post('/signup', (req, res) => {
   try {
-    fs.writeFile('users.json', JSON.stringify(req.body), err => {
-      if (err) {
-        res.status(401).json(err);
-        return;
-      }
-      res.status(201).json(req.body);
-    });
+    if (fs.existsSync('users.json')) {
+      fs.readFile('users.json', (err, users) => {
+        if (err) {
+          res.status(401).json(err);
+          return;
+        }
+        users = JSON.parse(`[${users}]`);
+        let user = users.find(one => one.email === req.body.email);
+        if (user) {
+          res.status(201).json({
+            message: 'user exists already', 
+            email: user.email,
+            username: user.username
+          });
+          return;
+        } else {
+          fs.appendFile('users.json', `,${JSON.stringify(req.body)}`, err => {
+            if (err) {
+              res.status(401).json(err);
+              return;
+            }
+            res.status(201).json({
+              message: 'user created successfully', 
+              email: req.body.email,
+              username: req.body.username
+            });
+          });
+        }
+      });
+    } else { 
+      fs.writeFile('users.json', JSON.stringify(req.body), err => {
+        if (err) {
+          res.status(401).json(err);
+          return;
+        }
+        res.status(201).json({
+          message: 'user created successfully', 
+          email: req.body.email,
+          username: req.body.username
+        });
+      });
+    }
   } catch(error) {
     res.status(401).json(error);
   }
@@ -23,17 +58,28 @@ app.post('/signup', (req, res) => {
 
 app.post('/login', (req, res) => {
   try {
-    fs.readFile('users.json', (err, user) => {
-      if (err) {
-        res.status(401).json(err);
-        return;
-      }
-      const data = {
-        signup: JSON.parse(user),
-        login: req.body
-      }
-      res.status(201).json(data);
-    });
+    if (fs.existsSync('users.json')) { 
+      fs.readFile('users.json', (err, users) => {
+        if (err) {
+          res.status(401).json(err);
+          return;
+        }
+        users = JSON.parse(`[${users}]`);
+        let user = users.find(one => one.email === req.body.email);
+        if (user) {
+          res.status(201).json({
+            message: 'login successful',
+            email: user.email,
+            username: user.username
+          });
+          return;
+        } else {
+          res.status(201).json({message: 'no user found, sign up instead'});
+        } 
+      });
+    } else { 
+      res.status(201).json({message: 'no user found, sign up instead'});
+    }
   } catch(error) {
     res.status(401).json(error);
   }
